WIP converting ruc API state gubbins to Open-Meteo
