+++ Skewt/State/SoundingState.swift
@@ -16,6 +16,7 @@ struct SoundingSelection: Codable, Hashable, Identifiable {
     }
     
     enum ModelType: Codable, CaseIterable, Identifiable, Equatable {
+        case automatic  // Allow data provider to select the best forecast model
         case op40
         case raob
         
@@ -42,6 +43,7 @@ struct SoundingSelection: Codable, Hashable, Identifiable {
     let type: ModelType
     let location: Location
     let time: Time
+    let dataAgeBeforeRefresh: TimeInterval
     
     var id: Self { self }
 }
@@ -52,6 +54,9 @@ extension SoundingSelection {
         type = .op40
         location = .closest
         time = .now
+        
+        let fiveMinutes = TimeInterval(5.0 * 60.0)
+        dataAgeBeforeRefresh = fiveMinutes
     }
 }
 
@@ -77,7 +82,7 @@ extension Date {
         let timeInterval = timeIntervalSince(referenceDate)
 
         switch modelType {
-        case .op40:
+        case .automatic, .op40:
             return timeInterval == 0.0 ? .now : .relative(timeInterval)
         case .raob:
             let intervalCount = Int(round(timeInterval / 60.0 / 60.0 / Double(modelType.hourInterval)))
@@ -113,6 +118,8 @@ extension SoundingSelection.Location {
 extension SoundingSelection.ModelType {
     var hourInterval: Int {
         switch self {
+        case .automatic:
+            return 1
         case .op40:
             return 1
 //        case .nam:
@@ -134,16 +141,17 @@ extension SoundingSelection {
         
         switch action {
         case .selectModelType(let type, let time):
-            return SoundingSelection(type: type, location: state.location, time: time)
+            return SoundingSelection(type: type, location: state.location, time: time, dataAgeBeforeRefresh: state.dataAgeBeforeRefresh)
         case .selectLocation(let location, let time):
-            return SoundingSelection(type: state.type, location: location, time: time)
+            return SoundingSelection(type: state.type, location: location, time: time, dataAgeBeforeRefresh: state.dataAgeBeforeRefresh)
         case .selectTime(let time):
-            return SoundingSelection(type: state.type, location: state.location, time: time)
+            return SoundingSelection(type: state.type, location: state.location, time: time, dataAgeBeforeRefresh: state.dataAgeBeforeRefresh)
         case .selectModelTypeAndLocation(let type, let location, let time):
             return SoundingSelection(
                 type: type ?? state.type,
                 location: location ?? state.location,
-                time: time
+                time: time,
+                dataAgeBeforeRefresh: state.dataAgeBeforeRefresh
             )
         }
     }
@@ -162,21 +170,38 @@ struct SoundingState: Codable {
         case doRefresh
         case changeAndLoadSelection(SoundingSelection.Action)
         case didReceiveFailure(SoundingError)
-        case didReceiveResponse(RucSounding)
-        case awaitSoundingLocation
+        case didReceiveResponse(OpenMeteoSoundingList)
     }
     
     enum Status: Codable {
         case idle
-        case awaitingSoundingLocationData
         case loading
-        case done(RucSounding, Date)
-        case refreshing(RucSounding)
+        case done(OpenMeteoSoundingList)
+        case refreshing(OpenMeteoSoundingList)
         case failed(SoundingError)
     }
     
     var selection: SoundingSelection
     var status: Status
+    
+    var sounding: Sounding? {
+        switch status {
+        case .idle, .loading, .failed(_):
+            return nil
+        case .done(let soundingList), .refreshing(let soundingList):
+            switch selection.time {
+            case .now:
+                return soundingList.closestSounding()
+            case .relative(let interval):
+                return soundingList.closestSounding(toDate: .now.addingTimeInterval(interval))
+            case .specific(let date):
+                return soundingList.closestSounding(toDate: date)
+            case.numberOfSoundingsAgo(let agoCount):
+                let secondsAgo = TimeInterval(selection.type.hourInterval) * 60.0 * 60.0
+                return soundingList.closestSounding(toDate: .now.addingTimeInterval(-secondsAgo))
+            }
+        }
+    }
 }
 
 // Default initializer
@@ -195,24 +220,12 @@ extension SoundingState {
 extension SoundingState.Status {
     var isLoading: Bool {
         switch self {
-        case .loading, .refreshing(_), .awaitingSoundingLocationData:
+        case .loading, .refreshing(_):
             return true
-        case .idle, .done(_, _), .failed(_):
+        case .idle, .done(_), .failed(_):
             return false
         }
 	}
-    
-    private var staleAge: TimeInterval { 60.0 * 60.0 }  // one hour
-    
-    /// Was this data fetched so long ago that it's likely out of date?
-    var isStale: Bool {
-        switch self {
-        case .done(_, let fetchDate):
-            return -fetchDate.timeIntervalSinceNow >= staleAge
-        case .idle, .awaitingSoundingLocationData, .loading, .refreshing(_), .failed(_):
-            return false
-        }
-    }
 }
 
 extension SoundingState.Action: CustomStringConvertible {
@@ -224,10 +237,8 @@ extension SoundingState.Action: CustomStringConvertible {
             return "Changing selection and reloading: \(selection)"
         case .didReceiveFailure(let error):
             return "Failed to load sounding: \(error)"
-        case .didReceiveResponse(let sounding):
-            return "Received sounding with \(sounding.data.dataPoints.count) data points"
-        case .awaitSoundingLocation:
-            return "Waiting for sounding location data"
+        case .didReceiveResponse(let soundingList):
+            return "Received list of \(soundingList.data.count) sounding\(soundingList.data.count != 1 ? "s" : "")"
         }
     }
 }
@@ -244,11 +255,11 @@ extension SoundingState {
             var state = state
             
             switch state.status {
-            case .done(let sounding, _):
-                if state.status.isStale {
+            case .done(let soundingList):
+                if Date.now.timeIntervalSince(soundingList.fetchTime) > state.selection.dataAgeBeforeRefresh {
                     state.status = .loading
                 } else {
-                    state.status = .refreshing(sounding)
+                    state.status = .refreshing(soundingList)
                 }
             default:
                 state.status = .loading
@@ -265,14 +276,9 @@ extension SoundingState {
             state.status = .failed(error)
             
             return state
-        case .didReceiveResponse(let sounding):
+        case .didReceiveResponse(let soundingList):
             var state = state
-            state.status = .done(sounding, .now)
-            
-            return state
-        case .awaitSoundingLocation:
-            var state = state
-            state.status = .awaitingSoundingLocationData
+            state.status = .done(soundingList)
             
             return state
         }
